+++ packages/workbox-broadcast-update/src/responsesAreSame.ts
@@ -6,9 +6,9 @@
   https://opensource.org/licenses/MIT.
 */
 
-import {WorkboxError} from 'workbox-core/_private/WorkboxError.mjs';
-import {logger} from 'workbox-core/_private/logger.mjs';
-import './_version.mjs';
+import {WorkboxError} from 'workbox-core/_private/WorkboxError.js';
+import {logger} from 'workbox-core/_private/logger.js';
+import './_version.js';
 
 /**
  * Given two `Response's`, compares several header values to see if they are
