+++ packages/workbox-broadcast-update/src/utils/constants.ts
@@ -6,7 +6,7 @@
   https://opensource.org/licenses/MIT.
 */
 
-import '../_version.mjs';
+import '../_version.js';
 
 export const CACHE_UPDATED_MESSAGE_TYPE = 'CACHE_UPDATED';
 export const CACHE_UPDATED_MESSAGE_META = 'workbox-broadcast-update';
